+++ services/typed-common/collections/realm.ts
@@ -1,4 +1,5 @@
 import mongoose from 'mongoose';
+import * as bcrypt from 'bcrypt';
 import { CollectionTypes } from '@microrealestate/types';
 
 const RealmSchema = new mongoose.Schema<CollectionTypes.Realm>({
@@ -9,6 +10,16 @@ const RealmSchema = new mongoose.Schema<CollectionTypes.Realm>({
       email: String,
       role: String,
       registered: Boolean,
+    },
+  ],
+  applications: [
+    {
+      name: String,
+      role: String,
+      clientId: String,
+      clientSecret: String,
+      createdDate: Date,
+      expiryDate: Date,
     },
   ],
   addresses: [
@@ -78,6 +89,18 @@ const RealmSchema = new mongoose.Schema<CollectionTypes.Realm>({
   },
   locale: String,
   currency: String,
+});
+//
+// hash application secrets before saving into database
+RealmSchema.pre('save', function (next) {
+  for (const app of this.applications) {
+    // chick if first save to hash secret
+    if (!app.createdDate) {
+      app.createdDate = new Date();
+      app.clientSecret = bcrypt.hashSync(app.clientSecret, 10);
+    }
+  }
+  next();
 });
 
 export default mongoose.model<CollectionTypes.Realm>('Realm', RealmSchema);