+++ types/common/collections.ts
@@ -24,6 +24,15 @@ export namespace CollectionTypes {
     createdDate?: Date;
   };
 
+  export type ApplicationCredentials = {
+    name: string;
+    role: string;
+    clientId: string;
+    clientSecret: string;
+    createdDate: Date;
+    expiryDate: Date;
+  };
+
   export type Realm = {
     _id: string;
     name: string;
@@ -33,6 +42,7 @@ export namespace CollectionTypes {
       role: string;
       registered: boolean;
     }[];
+    applications: CollectionTypes.ApplicationCredentials[];
     addresses: CollectionTypes.PartAddress[];
     bankInfo: {
       name: string;